requirements.txt
flask
opencv-python
aiohttp
pyserial
picamera2
ultralytics